Calculate beta and eta for non-orthogonal grids
Also return the radial grid spacing, store 'hrad' and 'hrad_ylow' for
later use.

To fix sign of vector normal to psi contours in calcBeta: need to
multiply f_R and f_Z by sign_bp because we want the radially-outward
vector perpendicular to psi contours, and if bpsign is negative then psi
increases inward instead of outward so (f_R,f_Z) would be in the
opposite direction.