Copy over positions at boundary locations from upper neighbour
Makes the positions at boundaries between regions consistent, when they
were taken from different MeshContour objects.

Also make all the corner positions at X-points have the value of the
X-point, not the nearest value in the MeshContours which had to start
near but not quite at the X-point in order for the ODE integrator to
know which direction to follow the psi-gradient in.