Put calculation of psi_vals into Equilibrium derived classes
Implementations of 'Equilibrium' must provide lists of psi-values to
their EquilibriumRegion objects, which then pass them to MeshRegion.
This was changed because it makes more sense to create the psi-values
for different regions at a point that knows about the global topology.
For example, for a disconnected-double-null equilibrium, we probably
want dpsidi constant between the two separatrices.

The base Equilibrium class does now provide a method to make
constructing psi(i) functions simple (assuming linear, quadratic or
cubic functions determined by the values and gradients at the end-points
are wanted).