Fix bug with creation of spacing functions in loop
Variables in Python lambdas are evaluated when the lambda is called, and
get values from variables in the scope where the lambda was created, but
these can change after the lambda is created. Work around this by
creating the lambda in the scope of a function whose local variables
don't change after it returns.