Fixes for getPolynomialPoloidalDistanceFunc
If a_lower is None when b_lower is not None, set a_lower to 0.;
if a_upper is None when b_upper is not None, set a_upper to 0.;
otherwise get unwanted sqrt contribution to spacing function at
lower/upper boundary.

Also fix the checks for the case when the polynomial part is supposed to
have zero gradient, with a tolerance for rounding errors giving a small
negative value.