Add option to fudge Bp.ylow near X-points
As Bp goes to zero at X-points, Bp.ylow can be very small even compared
to Bp.centre, leading to very large metric coefficients on the ylow
grid. As these can cause numerical problems, add an option which if
switched on (default is False) sets Bp.ylow at the points closest to
X-points to the average of the adjacent points in Bp.centre.