Fix bug with upper_intersect_index and behaviour of PsiContour.insert
When contours had to be extended to find the wall, so that
upper_intersect_index is negative, previously the intersection point was
being inserted into the contour as the last point, rather than the
second-to-last point where it should have been.

The behaviour of PsiContour.insert() was previously not consistent with
Python list's insert() method, which it is supposed to be. If a negative
index was passed the item was inserted one point off where it should
have been - e.g. insert(-1, item) should put the item one place before
the end of the list (replacing the item at index -1), not at the end of
the list.