Don't initialise Jcheck._corners when metrics do not have _corners
Previously the fix-up to make Jcheck._corners the same as J at X-points
(so the check does not fail just because J diverges where Bp goes to zero
at the X-point) initialised Jcheck even when its _corners field was
previously None. This made the check on J fail incorrectly. Now skip the
fix if Jcheck._corners is not initialised.