Replace polynomial spacing function with a monotonic function
Replace getPolynomialPoloidalDistanceFunc() with
getMonotonicPoloidalDistanceFunc() which is guaranteed always to return
a monotonic function, although it does not impose the constraint of zero
second derivative at the end points that
getPolynomialPoloidalDistanceFunc() did.

Note - getMonotonicPoloidalDistanceFunc() requires d_lower and d_upper
parameters always. In the unit tests which require
getMonotonicPoloidalDistanceFunc(), set parameters so that d_lower and
d_upper are passed with values that make the function linear, mimicing
the behaviour of getPolynomialPoloidalDistanceFunc() when d_lower and
d_upper were not passed.