Reset distance when reversing PsiContour
Ensures the distances are calculated consistent with the reversed
fine_contour. The values in distance are relative to the first point in
fine_contour, so the previous algorithm for reversing was wrong because
it assumed distances were relative to the first point in the PsiContour,
which might not be the same point.