Add pressure profile handling
If reading a GEQDSK file, the pressure is passed to
TokamakEquilibrium. This can optionally be extended with an
exponential profile outside the separatrix (using value and gradient
at the separatrix).

If a pressure is given to TokamakEquilibrium, then it is passed into
the EquilibriumRegion objects, adjusting for the PF regions by
reflecting the pressure profile in the leg about the separatrix.
This pressure is then output in the mesh as a 2D array "pressure".

Also added an example input YAML file for double null GEQDSK files.