Different gui-* installation options for different Qt wrappers
Instead of "pip install hypnotoad[gui]", have different versions for
pyside2 or PyQt5 (which the gui needs one but not both of, and in
principle we could use pyside or PyQt4 as well, if there was demand for
them). So now do "pip install hypnotoad[gui-pyside2]" to bring in
pyside2 or "pip install hypnotoad[gui-PyQt5]" to bring in PyQt5>=5.12.