Delete mesh when loading new geqdsk file
If a self.mesh object has already been created in the gui, delete it
when loading a new geqdsk file, since we create a new self.eq, so the
existing self.mesh is invalid.

Also move statements out of the try/except block in read_geqdsk() that
do not need to be in there.