Keep using defaults, store HypnotoadGui options in dict
Store the HypnotoadGui options in a dict, which contains only values
read from an input file or set by the user. This means that defaults can
be kept, and regenerated when the equilibrium or mesh objects are
recreated, since values changed by defaults in the internal options of
the equilibrium or mesh are not stored.