Improve parameters of orthogonal spacing function for nonorthogonal grids
Previously, the spacings at target and X-point used to create the
equilibriumRegion objects, which set the grid point positions that then
define the 'orthogonal' spacing functions, were set using
"nonorthogonal_xpoint_poloidal_spacing_length" and
"nonorthogonal_target_poloidal_spacing_length", which are the same parameters
as are used for the fixed perpendicular or poloidal spacing (which is
combined with the orthogonal spacing to make the final grid point
positions). There are two issues with this:
1. The orthogonal spacing functions cannot be altered after the Mesh
   object is first created (without some significant changes to the
   code). This means that if the nonorthogonal settings are changed (by
   using Mesh.redistributePoints()), the orthogonal spacing functions do
   not respect those changes.
2. The orthogonal spacing function and the fixed perpendicular or
   poloidal spacing function have to use the same spacing parameters
   (unless the user does some horrible hack either with the gui or a
   script to create the grid to set the orthogonal spacing parameters,
   then change the parameters and use redistributePoints() - that this is
   even possible is a bug because the final grid should be the same as if
   it was created by running an input file with the current settings).

The resolution to this is that when gridding the equilibriumRegion
objects during Mesh creation, the settings
"target_poloidal_spacing_length" and "xpoint_poloidal_spacing_length"
are used instead of "nonorthogonal_target_poloidal_spacing_length" and
"nonorthogonal_xpoint_poloidal_spacing_length". These parameters cannot
be changed after creation of the Mesh, which is a sort of resolution for
issue 1. By default "nonorthogonal_target_poloidal_spacing_length" is
the same as "target_poloidal_spacing_length" and
"nonorthogonal_xpoint_poloidal_spacing_length" is the same as
"xpoint_poloidal_spacing_length", but they can be set differently if
desired, allowing control over the fixed perpendicular or poloidal
spacing independent of the orthogonal spacing, fixing issue 2.