Use linear instead of cubic interpolation for FineContour.interpSSperp()
Linear interpolation avoids overshoots due to non-smooth s_perp when the
work-around to make a modified, monotonic s_perp is needed.
interpSSperp() is only used to place grid points, not to calculate
geometrical outputs, so the possible slight loss of accuracy should not
be a problem. If it is, accuracy can be increased by increasing
finecontour_Nfine.