Make grouping of regions deterministic.
Ensures zShift is consistent. Previously was using a set() to build the
groups, but the ordering of items it a set() is non-deterministic so the
groups for periodic parts of the domain (i.e. the core) could start with
different region in different runs if code or parameters were changed.