Replace contour point with wall point if they are too close
Previously, check was whether wall point was within atol of contour
point, and contour point was kept if so. This could cause errors due to
non-monotonic distance calculated on the contour if the atol was very
small. Since when a wall point is added, the contour points are about to
be redistributed, it is not important whether the contour point is kept
or not, it is better just to replace the contour point with the wall
point if they are close (to a loose tolerance).