Fix upper_intersect_index when lower wall point was a replacement
This bug should not have affected any current cases, because no regions
have walls at both ends.

The upper_intersect_index was previously always increased by 1 if a
lower wall was present. However, if the lower wall was close to a
contour point, the contour point is replaced by the wall point so the
number of points in the contour does not change - in this case the
upper_intersect_index does not need to be increased by 1.