Fix shifting of sfunc_orthogonal
When a wall point is added at the start of a contour, the function
represented by sfunc_orthogonal needs to be adjusted so that its value
at 'endInd' is the same as contour.totalDistance(), calculated for the
updated contour. Previously this was done in a way that might sometimes
be inaccurate (less important when finecontour_Nfine is large).