Fix circular geometry with non-constant q
Previously, q(r) was given as a polynomial with linear terms. The
analytical expressions given for non-constant q (i.e. more than one term
in the expansion) gave sensible answers for very restricted values of
the coefficients if at all. To be consistent at the magnetic axis, the
series for q(r) should only have even powers of r. The update provides a
correct expression for the new 2-coefficient case `q(r)=a0+a1*r**2`. It
does not seem to be possible to get a closed form expression for the
3-coefficient case `q(r)=a0+a1*r**2+a2*r**4` from Mathematica without it
being written as a complex expression, even though the value is real, so
this case is omitted.