Modify sys.path to fix integrated tests
`integrated_tests` is not registered as a package, so need to modify
sys.path. This is slightly hacky, but we don't want to install the
integrated tests as they include binary files.

Another option is to convert to pytest tests and rename `utils` to
`conftest` so it's automatically imported, but that seems more
complicated...