Fix changing value of string options
hypnotoad tries to parse the values set in the options table in the GUI.
This is needed to set floats, ints, etc. but often (always?) fails if
the value should have been a string. Catch some errors and just pass the
value as a string if they occur. Allows actually setting string options
in the GUI.