Ignore 'linewidths' or 'colors' arguments when they are not useful
`pyplot.scatter()` and `pyplot.plot()` error if they are passed
unrecognised keyword arguments. In `plotSeparatrix()` it is useful to
pass 'linewidths' and 'colors' in some cases (because sometimes
`pyplot.contour()` can be used to plot the separatrices), but they are
not valid arguments to `pyplot.scatter()` and `pyplot.plot()`, so delete
them from the kwargs if they are present.