Add follow_perpendicular_maxits option
The solve_ivp function used in follow_perpendicular has no maximum
number of iterations setting, so gets stuck on difficult meshes.

This adds a maximum number of iterations setting, defaulting to 1000.
If the solve fails, then it changes the psi locations of the points.
The result is that the user sees a grid rather than nothing, though
the grid points won't follow flux surfaces where the grid was modified.