Set the dimension for R_closed_wall and Z_closed_wall to 'closed_wall'
xBOUT detects grid files by checking that they do not have a "t"
dimension. However, boututils.Datafile defaults to writing a 1d array
with a "t" dimension, so the "R_closed_wall" and "Z_closed_wall"
variables were previously written with a "t" dimension, which made xBOUT
try to load the grid files as if they were dump files (which fails). Fix
this by specifying the name of the dimension as "closed_wall".