Add rounding-error tolerances to fix test failures under Python-3.13
Something (presumably changed rounding errors) was making some
conditionals evaluate to the wrong result for a separatrix segment
around the core. For these segments, the first and last points of the
FineContour and PsiContour are in the same place, so the FineContour
should not need to be extended, but without the tolerance introduced in
this commit, one was being extended, causing errors in
poloidal_distance_xlow.

Also increase tolerances in 'with rounding errors' test, as Python-3.13
has changed things slightly (maybe different random-number generation?).